Implement Pool 'value-changed' event callback args
Setting a new value and emitting the 'value-changed'
event has been delegated to the private function changeValue. This
new function accepts one parameter, the new value, records the
current value of the Pool object, and then sets the current value
equal to the new value. It then emits the 'value-changed' event,
and passes both the previous value and the new value to any
event handlers.